Beer bottle attack works now
The beer bottle now correctly explodes if it hits the player, and if it does so it will do 10 damage to them. It still explodes when it reaches the max range like previously. To accomplish this I changed the node to an Area2d(which is just a subclass of what it was before).

The puddle now also does damage per second to the player if it is in the puddle. The lifetime of the puddle, damage per second and tick interval are exported variables and adjustable.