collision rework
Ok so here are the collision layers

1 - Player
2 - Enemy
3 - Walls
4 - Player projectile(not used rn)
5 - Enemy projectile
8 - Intangible(the player while rolling)

So every enemy has a mask of 1 and 3(can detect walls and the player), every enemy projectile has 1 and sometimes 3 if it wants to detect walls, the player has 2 and 3. And walls should have everything

Since the bossdaddy and the player can actually collide with each other now we need to add special collision logic or else the boss just pushes around the player via move_and_slide(). I will do that later

Cool feature that the new layer system allows: now the beer bottle explodes when it hits a wall, not just when it hits the player. Small logic change to allow that